This is the initial commit of my portfolio website
I have created the initials and have put in only the essentials in this for now, and am planning to add more features to this website, like having a toggle for the weather to appear in the background of where I am right now. I will also try to implement things/ take inspirations from other portfolio websites.